Fix(set_game_speed) : increase game speed => level | Fix(game_play) : display and move piece after a limit of time

+++ src/prepare/game_init.c
@@ -8,7 +8,7 @@
 #include "tetris.h"
 
 extern const char BACKG_CHAR;
-extern int REFRESH_TIME;
+extern float REFRESH_TIME;
 
 static void game_init_struct(game_t *tetris)
 {
